refactor to module, use the clone method for IE
